+++ Cargo.toml
@@ -10,7 +10,6 @@ members = [
   "examples/rondpoint",
   "examples/sprites",
   "examples/todolist",
-  "examples/threadsafe",
   "fixtures/coverall",
   "fixtures/regressions/enum-without-i32-helpers",
   "fixtures/regressions/cdylib-crate-type-dependency/ffi-crate",
