+++ docs/manual/src/internals/object_references.md
@@ -10,21 +10,35 @@ including:
   active at any point in the program.
 * Guarding against data races.
 
-UniFFI aims to maintain these guarantees even when the Rust code is being invoked
-from a foreign language, at the cost of turning them into run-time checks rather
-than compile-time guarantees.
+The very nature of the problems UniFFI tries to solve is that calls may come
+from foreign languages on any thread, outside of the control of Rust's ownership
+system. UniFFI itself tries to take a hands-off approach as much as possible and
+depends on the Rust compiler itself to uphold safety guarantees, without assuming
+that foreign-language callers will be "well behaved".
+
+UniFFI's hands-off approach means that all object instances exposed by UniFFI must be safe to
+access concurrently. In Rust terminology, they must be `Send+Sync` and must be useable
+without taking any `&mut` references.
+
+Typically this will mean that the Rust implementation of an object uses some of Rust's
+data structures for thread-safe interior mutability, such as a `Mutex` or `RwLock` or
+the types from `std::atomic`. The precise details are completely up to the author
+of the component - as much as possible, UniFFI tries to stay out of your way, simply requiring
+that the object implementation is `Send+Sync` and letting the Rust compiler ensure that
+this is so.
 
 ## Handle Maps
 
-We achieve this by indirecting all object access through a
-[handle map](https://docs.rs/ffi-support/0.4.0/ffi_support/handle_map/index.html),
-a mapping from opaque integer handles to object instances. This indirection
+For additional typechecking safety, UniFFI indirects all object access through a
+"handle map", a mapping from opaque integer handles to object instances. This indirection
 imposes a small runtime cost but helps us guard against errors or oversights
 in the generated bindings.
 
 For each interface declared in the UDL, the UniFFI-generated Rust scaffolding
 will create a global handlemap that is responsible for owning all instances
 of that interface, and handing out references to them when methods are called.
+The handlemap requires that its contents be `Send+Sync`, helping enforce requirements
+around thread-safety.
 
 For example, given a interface definition like this:
 
@@ -40,7 +54,7 @@ The Rust scaffolding would define a lazyily-initialized global static like:
 
 ```rust
 lazy_static! {
-    static ref UNIFFI_HANDLE_MAP_TODOLIST: ConcurrentHandleMap<TodoList> = ConcurrentHandleMap::new();
+    static ref UNIFFI_HANDLE_MAP_TODOLIST: ArcHandleMap<TodoList> = ArcHandleMap::new();
 }
 ```
 
@@ -57,7 +71,7 @@ pub extern "C" fn todolist_TodoList_new(err: &mut ExternError) -> u64 {
 ```
 
 When invoking a method on the instance, the foreign-language code passes the integer handle back
-to the Rust code, which borrows a mutable reference to the instance from the handlemap for the duration
+to the Rust code, which borrows a reference to the instance from the handlemap for the duration
 of the method call:
 
 ```rust
@@ -85,21 +99,4 @@ This indirection gives us some important safety properties:
   then the handlemap will throw an error with high probability, providing some amount of run-time typechecking
   for correctness of the generated bindings.
 * The handlemap can ensure we uphold Rust's requirements around unique mutable references and threadsafey,
-  using a combination of compile-time checks and runtime locking depending on the details of the underlying
-  Rust struct that implements the interface.
-
-## Managing Concurrency
-
-By default, UniFFI uses the [ffi_support::ConcurrentHandleMap](https://docs.rs/ffi-support/0.4.0/ffi_support/handle_map/struct.ConcurrentHandleMap.html) struct as the handlemap for each declared instance. This class
-wraps each instance with a `Mutex`, which serializes access to the instance and upholds Rust's guarantees
-against shared mutable access.  This approach is simple and safe, but it means that all method calls
-on an instance are run in a strictly sequential fashion, limiting concurrency.
-
-For instances that are explicited tagged with the `[Threadsafe]` attribute, UniFFI instead uses
-a custom `ArcHandleMap` struct. This replaces the run-time `Mutex` with compile-time assertions
-about the safety of the underlying Rust struct. Specifically:
-
-* The `ArcHandleMap` will never give out a mutable reference to an instance, forcing the
-  underlying struct to use interior mutability and manage its own locking.
-* The `ArcHandleMap` can only contain structs that are `Sync` and `Send`, ensuring that
-  shared references can safely be accessed from multiple threads.
+  by specifying that the contained type must be `Send+Sync`, and by refusing to hand out any mutable references.
