+++ examples/sprites/src/sprites.udl
@@ -13,7 +13,6 @@ dictionary Vector {
   double dy;
 };
 
-[Threadsafe]
 interface Sprite {
   constructor(Point? initial_position);
   [Name=new_relative_to] constructor(Point reference, Vector direction);
