--- examples/threadsafe/src/threadsafe.udl
@@ -1,16 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-namespace threadsafe {};
-
-interface Counter {
-  void busy_wait(i32 ms);
-  i32 increment_if_busy();
-};
-
-[Threadsafe]
-interface ThreadsafeCounter {
-  void busy_wait(i32 ms);
-  i32 increment_if_busy();
-};
-
