--- examples/threadsafe/tests/test_generated_bindings.rs
@@ -1,8 +1,0 @@
-uniffi_macros::build_foreign_language_testcases!(
-    "src/threadsafe.udl",
-    [
-        // "tests/bindings/test_sprites.py",
-        "tests/bindings/test_threadsafe.kts",
-        // "tests/bindings/test_threadsafe.swift",
-    ]
-);
