+++ examples/todolist/src/todolist.udl
@@ -12,7 +12,6 @@ enum TodoError {
     "TodoDoesNotExist", "EmptyTodoList", "DuplicateTodo", "EmptyString", "DeligatedError"
 };
 
-[Threadsafe]
 interface TodoList {
     constructor();
     [Throws=TodoError]
